Dynamic Installer
Working on the dyamic install process.

Saving progress. Working on how to dynamically do a device
discovery. The dynamic-auto-install.sh is more of a discovery
script that finds the fastest non-removable disk and the on-board
ethernet to use for OS and MGMT.

Just transfered code from mac to linuxbox as automating and
messing with the ISO is easier with their tool
proxmox-auto-install-assistant for debian-based systems.