Refactoring
Using ENVs to make the configuration more flexible. Previously

The LAN IP address was preset without the option to change it
Added dev-env.sh file that can be sourced and will not overried
any exsting envs.

Adjusted the safety_config template to use envs for LAN address
and updated the insatller.

Set the default out-ot-band to be DHCP, can easily be overridden
in the ISO installer form and with ENV vars.

Updated the OPNsense installer to use an ENV for the password,
still has a default but this is highly discouraged.